ajustes no crud

+++ WEB/PROFESSOR/crudv2/src/components/crud/professor/ProfessorTableRow.jsx
@@ -3,12 +3,16 @@ import { Link } from "react-router-dom";
 import axios from "axios";
 
 const ProfessorTableRow = (props) => {
-    const {id,name,university,degree} = props.professor
+    const { id, name, university, degree } = props.professor
 
-    function deleteProfessor(){
-        axios.delete(`http://localhost:3001/professors/${id}`)
-        .then(response=>console.log("OK"))
-        .catch(error=>console.log(error))
+    function deleteProfessor() {
+        if (window.confirm(`Deseja excluir o elemento de ID: ${id}?`)) {
+            axios.delete(`http://localhost:3001/professors/${id}`)
+                .then(response => props.deleteProfessorById(id))
+                .catch(error => console.log(error))
+
+        }
+
     }
 
     return (
@@ -25,11 +29,11 @@ const ProfessorTableRow = (props) => {
             <td>
                 {degree}
             </td>
-            <td style={{textAlign:"center"}}>
+            <td style={{ textAlign: "center" }}>
                 <Link to={`/editProfessor/${id}`} className="btn btn-primary">Editar</Link>
             </td>
-            <td style={{textAlign:"center"}}>
-                <button className="btn btn-danger" onClick={()=>deleteProfessor()}>Apagar</button>
+            <td style={{ textAlign: "center" }}>
+                <button className="btn btn-danger" onClick={() => deleteProfessor()}>Apagar</button>
             </td>
         </tr>
     )
